+++ straightedge/diagrams/templates/__init__.py
@@ -17,6 +17,7 @@ from . import animated_trace  # noqa: F401
 from . import graph_traversal  # noqa: F401
 from . import graph_algorithm  # noqa: F401
 from . import havel_hakimi  # noqa: F401
+from . import floyd_warshall  # noqa: F401
 from . import disjoint_set  # noqa: F401
 from . import priority_queue  # noqa: F401
 from . import block_cut_tree  # noqa: F401
@@ -69,6 +70,7 @@ __all__ = [
     "graph_traversal",
     "graph_algorithm",
     "havel_hakimi",
+    "floyd_warshall",
     "disjoint_set",
     "priority_queue",
     "block_cut_tree",